Start factoring out generic tests
